Refactor setup methods
