+++ godtools/Views/TractElements/TractCallToAction.swift
@@ -63,7 +63,7 @@ class TractCallToAction: BaseTractElement {
     // MARK: - Setup
     
     override func setupView(properties: Dictionary<String, Any>) {
-        self.frame = buildFrame()
+        super.setupView(properties: properties)
         addArrowButton()
     }
     
@@ -93,9 +93,7 @@ class TractCallToAction: BaseTractElement {
         return textStyle
     }
     
-    // MARK: - Helpers
-    
-    fileprivate func buildFrame() -> CGRect {
+    override func buildFrame() -> CGRect {
         return CGRect(x: self.xPosition,
                       y: self.yPosition,
                       width: self.width,
