+++ godtools/Views/TractElements/TractHeader.swift
@@ -43,13 +43,11 @@ class TractHeader: BaseTractElement {
     }
     
     override func setupView(properties: [String: Any]) {
-        self.frame = buildFrame()
+        super.setupView(properties: properties)
         self.backgroundColor = self.primaryColor?.withAlphaComponent(0.9)
     }
     
-    // MARK: - Helpers
-    
-    fileprivate func buildFrame() -> CGRect {
+    override func buildFrame() -> CGRect {
         return CGRect(x: self.xPosition,
                       y: self.yPosition,
                       width: self.width,
