+++ godtools/Views/TractElements/TractHero.swift
@@ -31,13 +31,7 @@ class TractHero: BaseTractElement {
     
     // MARK: - Setup 
     
-    override func setupView(properties: Dictionary<String, Any>) {
-        self.frame = buildFrame()
-    }
-    
-    // MARK: - Helpers
-    
-    fileprivate func buildFrame() -> CGRect {
+    override func buildFrame() -> CGRect {
         return CGRect(x: self.xPosition,
                       y: self.yPosition,
                       width: self.width,
