+++ godtools/Views/TractElements/TractLabelGestures.swift
@@ -31,7 +31,7 @@ extension TractLabel {
     func handleGesture(sender: UITapGestureRecognizer) {
         if sender.state == .ended {
             let cardView = self.parent as! TractCard
-            cardView.didTapOnCard()
+            cardView.processCardWithState()
         }
     }
