+++ godtools/Views/TractElements/TractTab.swift
@@ -21,13 +21,7 @@ class TractTab: BaseTractElement {
     
     // MARK: - Setup
     
-    override func setupView(properties: [String: Any]) {
-        self.frame = buildFrame()
-    }
-    
-    // MARK: - Helpers
-    
-    fileprivate func buildFrame() -> CGRect {
+    override func buildFrame() -> CGRect {
         return CGRect(x: self.xPosition,
                       y: self.yPosition,
                       width: self.width,
