+++ godtools/Views/TractElements/TractTabs.swift
@@ -69,7 +69,7 @@ class TractTabs: BaseTractElement {
     }
     
     override func setupView(properties: [String: Any]) {
-        self.frame = buildFrame()
+        super.setupView(properties: properties)
         setupSegmentedControl()
     }
     
@@ -123,9 +123,7 @@ class TractTabs: BaseTractElement {
         self.elements = elements
     }
     
-    // MARK: - Helpers
-    
-    fileprivate func buildFrame() -> CGRect {
+    override func buildFrame() -> CGRect {
         return CGRect(x: self.xPosition,
                       y: self.yPosition,
                       width: self.width,
