+++ godtools/Views/TractElements/TractTitle.swift
@@ -40,10 +40,6 @@ class TractTitle: BaseTractElement {
     
     // MARK: - Setup
     
-    override func setupView(properties: Dictionary<String, Any>) {
-        self.frame = buildFrame()
-    }
-    
     override func textStyle() -> TractTextContentProperties {
         let textStyle = super.textStyle()
         textStyle.color = .gtWhite
@@ -60,9 +56,7 @@ class TractTitle: BaseTractElement {
         return textStyle
     }
     
-    // MARK: - Helpers
-    
-    fileprivate func buildFrame() -> CGRect {
+    override func buildFrame() -> CGRect {
         return CGRect(x: self.xPosition,
                       y: self.yPosition,
                       width: self.width,
